+++ src/one/nio/rpc/RpcPacket.java
@@ -18,13 +18,12 @@ package one.nio.rpc;
 
 import one.nio.net.Socket;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-
 import java.io.IOException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 class RpcPacket {
-    private static final Log log = LogFactory.getLog(RpcPacket.class);
+    private static final Logger log = LoggerFactory.getLogger(RpcPacket.class);
 
     private static final int WARN_PACKET_SIZE = 4 * 1024 * 1024;
     private static final int ERROR_PACKET_SIZE = 128 * 1024 * 1024;
@@ -59,7 +58,7 @@ class RpcPacket {
         if (size <= 0 || size >= ERROR_PACKET_SIZE) {
             throw new IOException("Invalid RPC packet from " + socket.getRemoteAddress());
         } else if (size >= WARN_PACKET_SIZE) {
-            log.warn("RPC packet from " + socket.getRemoteAddress() + " is too large: " + size);
+            log.warn("RPC packet from {} is too large: {}", socket.getRemoteAddress(), size);
         }
     }
 }