+++ src/one/nio/server/WorkerPool.java
@@ -17,8 +17,9 @@
 package one.nio.server;
 
 import one.nio.os.SchedulingPolicy;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.concurrent.SynchronousQueue;
 import java.util.concurrent.ThreadFactory;
@@ -27,7 +28,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
 final class WorkerPool extends ThreadPoolExecutor implements ThreadFactory, Thread.UncaughtExceptionHandler {
-    private static final Log log = LogFactory.getLog(WorkerPool.class);
+    private static final Logger log = LoggerFactory.getLogger(WorkerPool.class);
 
     private final AtomicInteger index;
     private final int threadPriority;
@@ -72,7 +73,7 @@ final class WorkerPool extends ThreadPoolExecutor implements ThreadFactory, Thre
 
     @Override
     public void uncaughtException(Thread t, Throwable e) {
-        log.error("Uncaught exception in " + t, e);
+        log.error("Uncaught exception in {}", t, e);
     }
 
     private static final class WaitingSynchronousQueue extends SynchronousQueue<Runnable> {